requirements.txt

streamlit
presidio-analyzer
python-dotenv
transformers
torch
